Created a new src folder with multiple files and classes inside
The src folder will hold all of the fundamental files that make the game of chess such as:
board
game
piece
square
and const
This allows for a rudimental display of the chess board initialized with square objects on each row/col. The next updates need to be make on the piece folder to flesh out the piece classes.