New methods and general cleanup
Organized some method/variable names and added more docstrings. Also added new methods to the piece folder to add pieces to the chess board. Currently when it is ran you will see a chess board, all white pieces centered at the bottom and all black pieces centered at the top.

The next update will be for the pieces on the board to move around.