Added New Class
Created new dragger class to allow the user to pick up a piece and drag it around the board. Currently you cannot place down the piece on a new square as by default it will return to its original position. As it stands, the piece being dragged around is not the smoothest experience and is still a choppy, but I will look for a way to fix this in a later update. Currently, I want to create a new class to allow pieces to be placed down in new positions.