Found bug that is causing performance issues, and updated other areas of code for efficiency
So, I discovered how to make the game fullscreen, but with this I found a bug that is causing performance issues. When the piece is being dragged around there are copies/afterimages of the piece being rendered and blited behind the board and around the board when the piece is dragged there. This will be the first thing I work on for the next update.