Created New Class and Methods
I created a new class, Move, to keep track of the initial and final squares where a piece will be. Then, I made a lot of courses within the squares file to check if the space was empty or if a team or rival occupied the area. Most importantly, I also created a static method to check whether or not the list of moves was within range of the chess board. I made a new class named show_moves to allow the user to display the potential actions and adjusted the main loop to use this method. Finally, and most importantly, I created a new technique called calc_moves with a bunch of sub-methods for each piece to calculate the legal moves of each piece. It is not fully fleshed out as the pawns need the en passant rule added, and the king needs the ability to short/long castles. But the basics of every piece move are coded in and will show legal actions.