Finished Product, But Needs Polishing
The game of chess, but certain checks and blocks need to be reworked for the game to fully be played as intended, but you can still play a full normalish game of chess. With all the rules and valid moves implemented, pieces can move, pieces can capture, pieces can block checks, etc.